requirements.txt
PyYAML>=6.0
psycopg[binary]>=3.2
requests>=2.31
pytest>=8.0
